+++ src/commonMain/kotlin/space/kscience/gdml/defines.kt
@@ -5,7 +5,6 @@ package space.kscience.gdml
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.UseSerializers
-import nl.adaptivity.xmlutil.serialization.XmlDefault
 
 @Serializable
 public sealed class GdmlDefine : GdmlNode
@@ -36,33 +35,30 @@ public class GdmlVariable(
 @Serializable
 @SerialName("position")
 public class GdmlPosition(
+    override var name: String,
     public var x: Number = 0f,
     public var y: Number = 0f,
     public var z: Number = 0f,
-    @XmlDefault("")
-    override var name: String = "",
     public var unit: LUnit? = null
 ) : GdmlDefine()
 
 @Serializable
 @SerialName("rotation")
 public class GdmlRotation(
+    override var name: String,
     public var x: Number = 0f,
     public var y: Number = 0f,
     public var z: Number = 0f,
-    @XmlDefault("")
-    override var name: String = "",
     public var unit: AUnit? = null
 ) : GdmlDefine()
 
 @Serializable
 @SerialName("scale")
 public class GdmlScale(
+    override var name: String,
     public var x: Number = 1.0,
     public var y: Number = 1.0,
     public var z: Number = 1.0,
-    @XmlDefault("")
-    override var name: String = ""
 ) : GdmlDefine()
 
 @Serializable