+++ src/commonMain/kotlin/space/kscience/gdml/solids.kt
@@ -223,7 +223,7 @@ public data class GdmlScaledSolid(
     override var name: String,
     @XmlSerialName("solidref", "", "")
     val solidref: GdmlRef<GdmlSolid>,
-    var scale: GdmlScale
+    var scale: GdmlScale,
 ) : GdmlSolid()
 
 /*
@@ -308,18 +308,20 @@ public sealed class GdmlBoolSolid : GdmlSolid() {
     public fun resolveFirstPosition(root: Gdml): GdmlPosition? = firstposition ?: firstpositionref?.resolve(root)
     public fun resolveFirstRotation(root: Gdml): GdmlRotation? = firstrotation ?: firstrotationref?.resolve(root)
 
-    @Deprecated("Use GdmlPosition constructor instead")
-    public fun position(x: Number = 0f, y: Number = 0f, z: Number = 0f): GdmlPosition = GdmlPosition().apply {
-        this.x = x
-        this.y = y
-        this.z = z
+    public fun position(x: Number = 0f, y: Number = 0f, z: Number = 0f) {
+        position = GdmlPosition("$name.position", x, y, z)
     }
 
-    @Deprecated("Use GdmlRotation constructor instead")
-    public fun rotation(x: Number = 0f, y: Number = 0f, z: Number = 0f): GdmlRotation = GdmlRotation().apply {
-        this.x = x
-        this.y = y
-        this.z = z
+    public fun rotation(x: Number = 0f, y: Number = 0f, z: Number = 0f) {
+        rotation = GdmlRotation("$name.rotation", x, y, z)
+    }
+
+    public fun firstposition(x: Number = 0f, y: Number = 0f, z: Number = 0f) {
+        firstposition = GdmlPosition("$name.firstPosition", x, y, z)
+    }
+
+    public fun firstrotation(x: Number = 0f, y: Number = 0f, z: Number = 0f) {
+        firstrotation = GdmlRotation("$name.firstRotation", x, y, z)
     }
 }